move nft pkg to netbase
- remove nft pkgs from both network/netlight pkgs
- move lansecurity rules template from cmds
