+++ pkg/netlight/resource/resource.go
@@ -14,9 +14,9 @@ import (
 	"github.com/threefoldtech/zosbase/pkg/netlight/bridge"
 	"github.com/threefoldtech/zosbase/pkg/netlight/ifaceutil"
 	"github.com/threefoldtech/zosbase/pkg/netlight/namespace"
-	"github.com/threefoldtech/zosbase/pkg/netlight/nft"
 	"github.com/threefoldtech/zosbase/pkg/netlight/options"
 	"github.com/threefoldtech/zosbase/pkg/netlight/tuntap"
+	"github.com/threefoldtech/zosbase/pkg/netbase/nft"
 	"github.com/threefoldtech/zosbase/pkg/zinit"
 	"github.com/vishvananda/netlink"
 )
