--- pkg/network/nft/nft.go
@@ -1,34 +1,0 @@
-package nft
-
-import (
-	"io"
-	"os/exec"
-
-	"github.com/rs/zerolog/log"
-
-	"github.com/pkg/errors"
-)
-
-// Apply applies the ntf configuration contained in the reader r
-// if ns is specified, the nft command is execute in the network namespace names ns
-func Apply(r io.Reader, ns string) error {
-	var cmd *exec.Cmd
-
-	if ns != "" {
-		cmd = exec.Command("ip", "netns", "exec", ns, "nft", "-f", "-")
-	} else {
-		cmd = exec.Command("nft", "-f", "-")
-	}
-
-	cmd.Stdin = r
-
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		log.Error().Err(err).Str("output", string(out)).Msg("error during nft")
-		if eerr, ok := err.(*exec.ExitError); ok {
-			return errors.Wrapf(err, "failed to execute nft: %v", string(eerr.Stderr))
-		}
-		return errors.Wrap(err, "failed to execute nft")
-	}
-	return nil
-}
